[feat/map_location] feat: refactor Location model and data source to improve package structure
