+++ feature/location-tracker/src/main/java/com/alxnophis/jetpack/location/tracker/data/repository/LocationRepositoryImpl.kt
@@ -2,13 +2,17 @@ package com.alxnophis.jetpack.location.tracker.data.repository
 
 import arrow.core.Either
 import com.alxnophis.jetpack.location.tracker.data.data.LocationDataSource
-import com.alxnophis.jetpack.location.tracker.domain.model.Location
-import com.alxnophis.jetpack.location.tracker.domain.model.LocationParameters
+import com.alxnophis.jetpack.location.tracker.data.model.Location
+import com.alxnophis.jetpack.location.tracker.data.model.LocationParameters
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.withContext
 
 internal class LocationRepositoryImpl(
     private val locationDataSource: LocationDataSource,
+    private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
 ) : LocationRepository {
     override val locationSharedFlow: SharedFlow<Location> =
         locationDataSource.locationSharedFlow
@@ -18,10 +22,14 @@ internal class LocationRepositoryImpl(
     override fun hasLocationAvailable(): Either<Unit, Unit> = locationDataSource.hasLocationAvailable()
 
     override suspend fun startLocationProvider(locationParameters: LocationParameters) {
-        locationDataSource.startLocationProvider(locationParameters)
+        withContext(ioDispatcher) {
+            locationDataSource.startLocationProvider(locationParameters)
+        }
     }
 
     override suspend fun stopLocationProvider() {
-        locationDataSource.stopLocationProvider()
+        withContext(ioDispatcher) {
+            locationDataSource.stopLocationProvider()
+        }
     }
 }
