--- feature/location-tracker/src/main/java/com/alxnophis/jetpack/location/tracker/domain/usecase/ProvideLastKnownLocationUseCase.kt
@@ -1,16 +1,0 @@
-package com.alxnophis.jetpack.location.tracker.domain.usecase
-
-import com.alxnophis.jetpack.location.tracker.data.repository.LocationRepository
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-
-class ProvideLastKnownLocationUseCase(
-    private val locationRepository: LocationRepository,
-    private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
-) {
-    suspend operator fun invoke() =
-        withContext(ioDispatcher) {
-            locationRepository.provideLastKnownLocationFlow()
-        }
-}
