--- feature/location-tracker/src/main/java/com/alxnophis/jetpack/location/tracker/domain/usecase/StartLocationProviderUseCase.kt
@@ -1,10 +1,0 @@
-package com.alxnophis.jetpack.location.tracker.domain.usecase
-
-import com.alxnophis.jetpack.location.tracker.data.repository.LocationRepository
-import com.alxnophis.jetpack.location.tracker.domain.model.LocationParameters
-
-class StartLocationProviderUseCase(
-    private val locationRepository: LocationRepository,
-) {
-    suspend operator fun invoke(locationParameters: LocationParameters = LocationParameters()) = locationRepository.startLocationProvider(locationParameters)
-}
